OnlinePackage

OnlinePackage.USER_FILE

OnlinePackage.delete

OnlinePackage.get_user

OnlinePackage.put_online

OnlinePackage.USER_FILE — Constant

where OnlinePackage will look for your user information

OnlinePackage.delete — Method

delete(user::User, repo_name)

Delete a repository.

OnlinePackage.get_user — Function

get_user(user_file = USER_FILE)

Create a user profile from a TOML file. By default, looks in USER_FILE

The file must contain a username, github_token, and ssh_keygen_file. See the sample.toml file in this repository for an example.

Use your github username.

Get a github_token here. Make sure to check the public_repo scope, and optionally, the delete_repo scope.

If ssh-keygen is in your path, just set ssh_keygen_file to "ssh-keygen". If not, it often comes prepacked with git; check PATH_TO_GIT/usr/bin/ssh-keygen".

OnlinePackage.put_online — Method

put_online(user::User, repo_name)

Put a repository online: put it on github, and create an SSH key to allow github actions to push to github. If the repository already exists, it will create a new set of keys.

julia> using OnlinePackage

julia> user = get_user(OnlinePackage.SAMPLE_FILE);

julia> put_online(user, "Test.jl")

julia> put_online(user, "Test.jl")

julia> delete(user, "Test.jl")